Clean code, documentation and update the demos.
