+++ src/domain/models/keyManagement/Curve.ts
@@ -1,0 +1,5 @@
+export enum Curve {
+  X25519 = "X25519",
+  ED25519 = "Ed25519",
+  SECP256K1 = "Secp256k1",
+}
