+++ src/domain/models/keyManagement/KeyCurve.ts
@@ -1,0 +1,6 @@
+import { Curve } from "./Curve";
+
+export interface KeyCurve {
+  curve: Curve;
+  index?: number;
+}
