+++ src/domain/models/keyManagement/KeyTypes.ts
@@ -1,0 +1,4 @@
+export enum KeyTypes {
+  "EC" = "EC",
+  "Curve25519" = "Curve25519",
+}
